Update frustum_or_perception_of_depth_pertaining_to_geometry.html
Intended version. There is a noticeable effect in comparison to with out the divide by z(+1). With out it, it appears that the back and front of the bottom of the pyramid when almost perfectly aligned are the exact same size when this divide by z is applied it’s clear that their is a noticeable difference in length, giving the perspective of distance/depth.